added colors & params selection
