+++ src/components/ParamsSelection/index.ts
@@ -1,0 +1,1 @@
+export * from './ParamsSelection';
